app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.5
h11==0.14.0
httptools==0.6.4
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
xmltodict==0.14.2
python-status==1.0.1
openai==1.54.4
flower==2.0.1
asyncpg==0.30.0
asgiref==3.8.1
pytest==8.3.3
httpx==0.27.2